added PriceDirection component
